requirements.txt
langchain
langchain-groq
langchain-community
streamlit
python-dotenv
